Fixed: Bug in maximum temperature setting of LR Scheduler.
Details:
The old implementation of the benchmark pipeline used the "Epoch" fidelity value as the maximum number of epochs that the LR scheduler will decay the LR for. This poses a problem when a learning curve needs to be extended beyond an initial setting, e.g. when attempting an incremental process for building a benchmark dataset, since the initial LR setting is not updated when the value of "Epoch" changes. In a correct implementation, the maximum tempurate/maximum epochs of the LR scheduler should be a part of the training configuration that remains consistent across all learning curves in a dataset. This has now been implemented as the parameter "sched_max_temp".

Caveats:
Learning curves with different values of "sched_max_temp" on the exact same configuration should be treated as completely different model trainings.