Update: Added new pipeline configuration option for target transformation.
The pipeline configuration can now include a set of transformations that are to be applied to the target data. These transformations are applied to the target data before the model is fit and the inverse transformation is applied to the predictions made by the model. Only a minimal set of transformations is supported right now. This functionality is still untested.