Fixed: Bug in update of XGBoost objective function.
Caught a bug in the implementation of the custom objective function pipeline that caused the model to not be aware of the updated objective function built using the specifications of the pipeline configuration. This also lead to catching an issue with the implementation of the exponential boundary losses - numerical overflows can occur when calculation the exponential loss and need to be addressed.