Fixed: Random sampling from performance dataset in the public API.
Indexing by configuration can cause the index of the DataFrame to become non-unique, since there are no guarantees that the same configuration was not sampled twice. Therefore, querying the performance dataset needs to be handled appropriately. For now, direct queries to the tabular dataset have been disabled in favour of allowing simple random sampling from it.